Restructure model validator
