+++ cg/services/order_validation_service/workflows/microsalt/validation_service.py
@@ -1,14 +1,13 @@
 from cg.services.order_validation_service.errors.order_errors import OrderError
 from cg.services.order_validation_service.errors.sample_errors import SampleError
 from cg.services.order_validation_service.errors.validation_errors import ValidationErrors
+from cg.services.order_validation_service.model_validator.model_validator import ModelValidator
 from cg.services.order_validation_service.order_validation_service import OrderValidationService
 from cg.services.order_validation_service.utils import (
     apply_order_validation,
     apply_sample_validation,
 )
-from cg.services.order_validation_service.workflows.microsalt.validation.field.model_validator import (
-    MicroSaltModelValidator,
-)
+from cg.services.order_validation_service.workflows.microsalt.models.order import MicrosaltOrder
 from cg.services.order_validation_service.workflows.microsalt.validation_rules import SAMPLE_RULES
 from cg.services.order_validation_service.response_mapper import create_order_validation_response
 from cg.services.order_validation_service.workflows.order_validation_rules import ORDER_RULES
@@ -25,7 +24,7 @@ class MicroSaltValidationService(OrderValidationService):
         return create_order_validation_response(raw_order=raw_order, errors=errors)
 
     def _get_errors(self, raw_order: dict) -> ValidationErrors:
-        order, field_errors = MicroSaltModelValidator.validate(raw_order)
+        order, field_errors = ModelValidator.validate(order=raw_order, model=MicrosaltOrder)
 
         if field_errors:
             return field_errors
