+++ cg/services/order_validation_service/workflows/mip_dna/validation_service.py
@@ -2,15 +2,14 @@ from cg.services.order_validation_service.errors.case_errors import CaseError
 from cg.services.order_validation_service.errors.case_sample_errors import CaseSampleError
 from cg.services.order_validation_service.errors.order_errors import OrderError
 from cg.services.order_validation_service.errors.validation_errors import ValidationErrors
+from cg.services.order_validation_service.model_validator.model_validator import ModelValidator
 from cg.services.order_validation_service.order_validation_service import OrderValidationService
 from cg.services.order_validation_service.utils import (
     apply_case_sample_validation,
     apply_case_validation,
     apply_order_validation,
 )
-from cg.services.order_validation_service.workflows.mip_dna.validation.field.model_validator import (
-    MipDnaModelValidator,
-)
+from cg.services.order_validation_service.workflows.mip_dna.models.order import MipDnaOrder
 from cg.services.order_validation_service.workflows.mip_dna.validation_rules import (
     CASE_RULES,
     CASE_SAMPLE_RULES,
@@ -30,7 +29,7 @@ class MipDnaValidationService(OrderValidationService):
         return create_order_validation_response(raw_order=raw_order, errors=errors)
 
     def _get_errors(self, raw_order: dict) -> ValidationErrors:
-        order, field_errors = MipDnaModelValidator.validate(raw_order)
+        order, field_errors = ModelValidator.validate(order=raw_order, model=MipDnaOrder)
 
         if not order:
             return field_errors
