+++ cg/services/order_validation_service/workflows/order_validation_rules.py
@@ -1,4 +1,4 @@
-from cg.services.order_validation_service.rules.order.order_rules import (
+from cg.services.order_validation_service.rules.order.rules import (
     validate_customer_can_skip_reception_control,
     validate_customer_exists,
     validate_ticket_number_required_if_connected,
