--- cg/services/order_validation_service/workflows/tomte/validation/field/tomte_model_validator.py
@@ -1,16 +1,0 @@
-from pydantic_core import ValidationError
-
-from cg.services.order_validation_service.errors.validation_errors import ValidationErrors
-from cg.services.order_validation_service.validators.field.utils import convert_errors
-from cg.services.order_validation_service.workflows.tomte.models.order import TomteOrder
-
-
-class TomteModelValidator:
-
-    @classmethod
-    def validate(cls, raw_order: dict) -> tuple[TomteOrder | None, ValidationErrors]:
-        try:
-            order = TomteOrder.model_validate(raw_order)
-            return order, ValidationErrors()
-        except ValidationError as error:
-            return None, convert_errors(pydantic_errors=error)
