+++ cg/services/order_validation_service/workflows/tomte/validation_service.py
@@ -2,6 +2,7 @@ from cg.services.order_validation_service.errors.case_errors import CaseError
 from cg.services.order_validation_service.errors.case_sample_errors import CaseSampleError
 from cg.services.order_validation_service.errors.order_errors import OrderError
 from cg.services.order_validation_service.errors.validation_errors import ValidationErrors
+from cg.services.order_validation_service.model_validator.model_validator import ModelValidator
 from cg.services.order_validation_service.order_validation_service import (
     OrderValidationService,
 )
@@ -12,9 +13,7 @@ from cg.services.order_validation_service.utils import (
 )
 from cg.services.order_validation_service.response_mapper import create_order_validation_response
 from cg.services.order_validation_service.workflows.order_validation_rules import ORDER_RULES
-from cg.services.order_validation_service.workflows.tomte.validation.field.tomte_model_validator import (
-    TomteModelValidator,
-)
+from cg.services.order_validation_service.workflows.tomte.models.order import TomteOrder
 from cg.services.order_validation_service.workflows.tomte.validation_rules import (
     CASE_RULES,
     CASE_SAMPLE_RULES,
@@ -32,7 +31,7 @@ class TomteValidationService(OrderValidationService):
         return create_order_validation_response(raw_order=raw_order, errors=errors)
 
     def _get_errors(self, raw_order: dict) -> ValidationErrors:
-        order, field_errors = TomteModelValidator.validate(raw_order)
+        order, field_errors = ModelValidator.validate(order=raw_order, model=TomteOrder)
 
         if not order:
             return field_errors
